requirements.txt
Flask
gunicorn
mysql-connector-python
flask-bcrypt
python-dotenv
PyJWT
scikit-learn
numpy
pandas